# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mini Compiler & Static Code Analyzer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SourceInput {
  /** Source code to compile */
  source: string;
  /**
   * Optional filename for display
   * @nullable
   */
  filename?: string | null;
}

export interface Token {
  /** Token type (KEYWORD, IDENTIFIER, NUMBER, STRING, OPERATOR, DELIMITER, COMMENT) */
  type: string;
  value: string;
  line: number;
  column: number;
}

export interface CompilerError {
  /** lexer | parser | semantic | optimizer */
  phase: string;
  message: string;
  line: number;
  /** @nullable */
  column?: number | null;
}

export interface TokenizeResult {
  tokens: Token[];
  errors: CompilerError[];
  tokenizationTimeMs: number;
  tokenCount: number;
}

export interface ASTNode {
  type: string;
  value: string;
  /** @nullable */
  line?: number | null;
  children?: ASTNode[];
}

export interface ParseResult {
  ast: ASTNode;
  errors: CompilerError[];
  parseTimeMs: number;
  nodeCount: number;
}

export interface SemanticWarning {
  /** Warning type (UNUSED_VARIABLE, UNDEFINED_VARIABLE, TYPE_MISMATCH, SCOPE_VIOLATION, etc.) */
  type: string;
  message: string;
  line: number;
  /** error | warning | info */
  severity?: string;
}

export interface StaticIssue {
  /** Issue type (DEAD_CODE, INFINITE_LOOP, UNUSED_FUNCTION, DUPLICATE_CONDITION, etc.) */
  type: string;
  message: string;
  line: number;
  severity?: string;
  /** @nullable */
  suggestion?: string | null;
}

export interface SymbolEntry {
  name: string;
  type: string;
  scope: string;
  line: number;
  used?: boolean;
}

export interface AnalysisResult {
  semanticWarnings: SemanticWarning[];
  staticIssues: StaticIssue[];
  /** 0-100 maintainability score */
  maintainabilityScore: number;
  analysisTimeMs: number;
  symbolTable: SymbolEntry[];
  complexityScore?: number;
}

export interface OptimizationPass {
  /** e.g. CONSTANT_FOLDING, DEAD_CODE_ELIMINATION */
  name: string;
  applied: boolean;
  description: string;
  instancesFound?: number;
  estimatedSpeedupPct?: number;
}

export interface OptimizeResult {
  passes: OptimizationPass[];
  optimizedSource: string;
  optimizationTimeMs: number;
  totalEstimatedSpeedupPct: number;
  memoryReductionPct?: number;
}

export interface IRInstruction {
  /** Three-address code operation (ASSIGN, ADD, SUB, MUL, DIV, JUMP, LABEL, CALL, etc.) */
  op: string;
  result: string;
  /** @nullable */
  arg1?: string | null;
  /** @nullable */
  arg2?: string | null;
  /** @nullable */
  block?: string | null;
}

export interface IRResult {
  instructions: IRInstruction[];
  blocks: string[];
  irGenerationTimeMs: number;
}

export interface PipelineStage {
  name: string;
  /** success | error | warning | skipped */
  status: string;
  timeMs: number;
  itemCount?: number;
}

export interface CompileResult {
  tokens: Token[];
  ast: ASTNode;
  analysis: AnalysisResult;
  optimization: OptimizeResult;
  ir: IRResult;
  pipeline: PipelineStage[];
  totalTimeMs: number;
  errors: CompilerError[];
  maintainabilityScore: number;
}

export interface SampleProgram {
  id: string;
  name: string;
  description: string;
  source: string;
  /** basic | loops | functions | advanced */
  category: string;
}

export type CompilerStatsTopIssueTypesItem = {
  type: string;
  count: number;
};

export interface CompilerStats {
  totalRuns: number;
  avgTokenizationMs: number;
  avgParseMs: number;
  avgOptimizationMs: number;
  avgMaintainabilityScore: number;
  topIssueTypes: CompilerStatsTopIssueTypesItem[];
}