# lib/api-zod/src/generated/types/aSTNode.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mini Compiler & Static Code Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface ASTNode {
  type: string;
  value: string;
  /** @nullable */
  line?: number | null;
  children?: ASTNode[];
}