# lib/api-zod/src/generated/types/iRInstruction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mini Compiler & Static Code Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface IRInstruction {
  /** Three-address code operation (ASSIGN, ADD, SUB, MUL, DIV, JUMP, LABEL, CALL, etc.) */
  op: string;
  result: string;
  /** @nullable */
  arg1?: string | null;
  /** @nullable */
  arg2?: string | null;
  /** @nullable */
  block?: string | null;
}