# lib/api-zod/src/generated/types/optimizeResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mini Compiler & Static Code Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { OptimizationPass } from "./optimizationPass";

export interface OptimizeResult {
  passes: OptimizationPass[];
  optimizedSource: string;
  optimizationTimeMs: number;
  totalEstimatedSpeedupPct: number;
  memoryReductionPct?: number;
}