# lib/api-zod/src/generated/types/parseResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mini Compiler & Static Code Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { ASTNode } from "./aSTNode";
import type { CompilerError } from "./compilerError";

export interface ParseResult {
  ast: ASTNode;
  errors: CompilerError[];
  parseTimeMs: number;
  nodeCount: number;
}