# lib/api-zod/src/generated/types/staticIssue.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mini Compiler & Static Code Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface StaticIssue {
  /** Issue type (DEAD_CODE, INFINITE_LOOP, UNUSED_FUNCTION, DUPLICATE_CONDITION, etc.) */
  type: string;
  message: string;
  line: number;
  severity?: string;
  /** @nullable */
  suggestion?: string | null;
}