# lib/api-zod/src/generated/types/token.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mini Compiler & Static Code Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface Token {
  /** Token type (KEYWORD, IDENTIFIER, NUMBER, STRING, OPERATOR, DELIMITER, COMMENT) */
  type: string;
  value: string;
  line: number;
  column: number;
}