# lib/api-zod/src/generated/types/tokenizeResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mini Compiler & Static Code Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { CompilerError } from "./compilerError";
import type { Token } from "./token";

export interface TokenizeResult {
  tokens: Token[];
  errors: CompilerError[];
  tokenizationTimeMs: number;
  tokenCount: number;
}